Add analytics support to SlashCommandListener
An AnalyticsRepository has been added to the SlashCommandListener constructor and used in the handle event method to track the usage statistics related to the usage of slash commands. This update will include changes such as the time of usage and frequency of specific slash command actions. This change provides more detailed analytics which will be helpful for optimizing the bot's interaction with users.